+++ stable-diffusion-abi.h
@@ -74,7 +74,7 @@ STABLE_DIFFUSION_API new_sd_ctx_go_params *new_sd_ctx_params(
         const char *model_path,
         const char *lora_model_dir,
         const char *vae_path,
-        int n_threads,
+        int16_t n_threads,
         enum sd_type_t wtype,
         enum rng_type_t rng_type,
         enum schedule_t schedule
